Move MBeanQuery out of package entity
Migrate https://github.com/netdata/netdata/pull/2663/commits/d300b958fbf59c470e3e193441aee0216a3b2519
